+++ openpyxl/writer/excel.py
@@ -28,7 +28,6 @@ from openpyxl.packaging.relationship import (
     Relationship,
 )
 from openpyxl.comments.comment_sheet import CommentSheet
-from openpyxl.packaging.extended import ExtendedProperties
 from openpyxl.styles.stylesheet import write_stylesheet
 from openpyxl.worksheet._writer import WorksheetWriter
 from openpyxl.workbook._writer import WorkbookWriter
@@ -52,6 +51,7 @@ class ExcelWriter(object):
 
 
     def write_data(self):
+        from openpyxl.packaging.extended import ExtendedProperties
         """Write the various xml files into the zip archive."""
         # cleanup all worksheets
         archive = self._archive